+++ src/main/java/shopping/domain/cart/Quantity.java
@@ -1,4 +1,4 @@
-package shopping.domain;
+package shopping.domain.cart;
 
 
 import shopping.exception.ErrorType;
