+++ src/main/java/shopping/repository/CartItemRepository.java
@@ -2,7 +2,7 @@ package shopping.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import shopping.domain.CartItem;
+import shopping.domain.cart.CartItem;
 
 import java.util.List;
