+++ src/main/java/shopping/repository/UserRepository.java
@@ -2,9 +2,9 @@ package shopping.repository;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shopping.domain.Email;
-import shopping.domain.Password;
-import shopping.domain.User;
+import shopping.domain.user.Email;
+import shopping.domain.user.Password;
+import shopping.domain.user.User;
 
 import java.util.Optional;
