+++ src/test/java/shopping/domain/cart/CartItemTest.java
@@ -1,7 +1,9 @@
-package shopping.domain;
+package shopping.domain.cart;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import shopping.domain.product.Product;
+import shopping.domain.user.User;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -11,8 +13,8 @@ class CartItemTest {
     @DisplayName("장바구니 항목 수량을 1 증가할 수 있다.")
     void increaseQuantity() {
         // given
-        Product product = new Product();
-        User user = new User();
+        User user = new User(1L, "test@test.com", "test");
+        Product product = new Product(1L, "치킨", "chicken.png", 2000);
         CartItem cartItem = new CartItem(user, product, 1);
 
         // when
@@ -26,8 +28,8 @@ class CartItemTest {
     @DisplayName("장바구니 항목 수량을 업데이트 할 수 있다.")
     void updateQuantity() {
         // given
-        Product product = new Product();
-        User user = new User();
+        User user = new User(1L, "test@test.com", "test");
+        Product product = new Product(1L, "치킨", "chicken.png", 2000);
         CartItem cartItem = new CartItem(user, product, 1);
 
         // when
