+++ src/test/java/shopping/domain/cart/CartItemsTest.java
@@ -1,7 +1,9 @@
-package shopping.domain;
+package shopping.domain.cart;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import shopping.domain.product.Product;
+import shopping.domain.user.User;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -14,8 +16,8 @@ class CartItemsTest {
     @DisplayName("장바구니에 상품을 추가한다.")
     void addCartItem() {
         // given
-        User user = new User();
-        Product product = new Product();
+        User user = new User(1L, "test@test.com", "test");
+        Product product = new Product(1L, "치킨", "chicken.png", 2000);
         CartItem item = new CartItem(user, product, 1);
         CartItems items = new CartItems(new ArrayList<>());
         // when
@@ -29,7 +31,7 @@ class CartItemsTest {
     @DisplayName("장바구니에 존재하는 상품을 추가하면 수량이 늘어난다.")
     void addCartItemDuplicate() {
         // given
-        User user = new User();
+        User user = new User(1L, "test@test.com", "test");
         Product product = new Product(1L, "치킨", "chicken.png", 2000);
         CartItem item = new CartItem(user, product, 1);
         CartItems items = new CartItems(new ArrayList<>());
@@ -47,7 +49,7 @@ class CartItemsTest {
     @DisplayName("장바구니에 상품이 존재하면 true를 반환한다.")
     void contains() {
         // given
-        User user = new User();
+        User user = new User(1L, "test@test.com", "test");
         Product product = new Product(1L, "치킨", "chicken.png", 2000);
         CartItem item = new CartItem(1L, user, product, 1);
         CartItems items = new CartItems(new ArrayList<>(List.of(item)));
@@ -58,10 +60,10 @@ class CartItemsTest {
     }
 
     @Test
-    @DisplayName("장바구니에 상품이 존재하면 false를 반환한다.")
+    @DisplayName("장바구니에 상품이 존재하지 않으면 false를 반환한다.")
     void notContains() {
         // given
-        User user = new User();
+        User user = new User(1L, "test@test.com", "test");
         Product product = new Product(1L, "치킨", "chicken.png", 2000);
         CartItem item = new CartItem(1L, user, product, 1);
         CartItems items = new CartItems(new ArrayList<>());
