+++ src/test/java/shopping/domain/cart/QuantityTest.java
@@ -1,4 +1,4 @@
-package shopping.domain;
+package shopping.domain.cart;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
