+++ src/test/java/shopping/domain/user/PasswordTest.java
@@ -1,4 +1,4 @@
-package shopping.domain;
+package shopping.domain.user;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
