+++ src/test/java/shopping/dto/ProductResponseTest.java
@@ -2,7 +2,7 @@ package shopping.dto;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import shopping.domain.Product;
+import shopping.domain.product.Product;
 
 import static org.assertj.core.api.Assertions.assertThat;
